chore(ui): added a page wrapper class
